requirements.txt
# API
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2

# ML
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2

# MLOps
mlflow==2.8.1
elasticsearch==8.11.0

# Utilities
pyyaml==6.0.1
joblib==1.3.2
python-multipart==0.0.6
requests==2.31.0
